chezmoi: fix symlinks

+++ home/run_onchange_after_04-create_symlinks.sh
@@ -6,13 +6,21 @@ echo ">> Creating symlinks..."
 # sensitive files
 ########################################################
 
-ln -sf ~/Dropbox ~/Cloud
-ln -sf ~/Cloud/Sync/krew ~/.krew
+if [ ! -d ~/Cloud ]; then
+  ln -sf ~/Dropbox ~/Cloud
+fi
+
+if [ ! -d ~/.krew ]; then
+  ln -sf ~/Cloud/Sync/krew ~/.krew
+fi
 
 if [ ! -d ~/.local/share/ptpython ]; then
   mkdir -p ~/.local/share/ptpython
 fi
-ln -sf ~/Cloud/Sync/ptpython_history ~/.local/share/ptpython/history
+
+if [ -f ~/.local/share/ptpython/history ]; then
+  ln -sf ~/Cloud/Sync/ptpython_history ~/.local/share/ptpython/history
+fi
 
 ########################################################
 # aftersleep script
